Templated api.bigpanda.io string into WEB_API_BASE_URL parameter that is filled by frontier.
When reading documentation in local envs or in OD's the url that appears in the documentation refers to api.bigpanda.io which is prod and incorrect in those cases.
Furthermore, in the event that we change our api 'api.bigpanda.io' to something else, all configuration's will break and this commit prevents that.
The files responsible for the parameter value are : guide.template.jade and integration.instructions.controller.js in frontier project.